Update deployment script

+++ install.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 set -e
 
+if [[ $EUID -ne 0 ]]; then
+   1>&2 echo "This script must be run as root" 
+   exit 1
+fi
+
 THIS_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 TARGET=/var/www/html
 
 cd $THIS_DIR/client
-rm -f dist/*
+rm -rf dist/*
 # npm install
 npm run build
 rm -rf $TARGET.old
